% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_truncated_normal}
\alias{plot_truncated_normal}
\title{Plot Truncated Normal Distribution}
\usage{
plot_truncated_normal(cov, lower, upper, y, plot_type = "scatter")
}
\arguments{
\item{cov}{Numeric. The covariance between the two variables.}

\item{lower}{Numeric vector of length 2. The lower bounds for the truncation.}

\item{upper}{Numeric vector of length 2. The upper bounds for the truncation.}

\item{y}{Numeric. The upper limit for the y-axis in the histogram plot.}

\item{plot_type}{Character. The type of plot to generate. Either "scatter" for
a scatter plot of the truncated data or "hist" for a histogram with a density curve.}
}
\description{
This function generates plots for a truncated bivariate normal distribution.
It can either create a scatter plot of the generated data or a histogram of
one of the variables with its corresponding density curve.
}
\details{
The function generates data from a truncated bivariate normal distribution
using the \code{rtmvnorm} function from the \code{tmvtnorm} package. It then plots the data
based on the specified plot type using \code{ggplot2}.
}
\examples{
plot_truncated_normal(cov = 0.5, lower = c(-2, -2), upper = c(2, 2), y = 0.5, plot_type = "scatter")
plot_truncated_normal(cov = 0.3, lower = c(-1, -1), upper = c(1, 1), y = 0.5, plot_type = "hist")

}
